Added a few sentences to the documentation of the new merging stuff.
I think this stuff (the new merge implementation as well as, more
importantly, the new implementation of backups) is now just about
ready to be used by others.  In fact, it would be useful to have a few
more people try it to flush out any remaining bugs.  

Please keep a careful eye on what it is doing, though, if you choose
to upgrade.




git-svn-id: https://webdav.seas.upenn.edu/svn/unison/trunk@90 4bce34ff-96ee-0310-b826-95abcfff6e26